30(main) 리팩토링, 기능개선(검색 옵션 패널, useContext)
- useContext 적용 selectedTabName, optionTabUrl 상태 관리 간소화
